# openapi.yaml
openapi: 3.0.0
info:
  title: POS System API
  version: 1.0.0
  description: Point of Sale System API Documentation

servers:
  - url: http://localhost:8080/api/v1
    description: Local development server

paths:
  /auth/login:
    post:
      summary: User login
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                      username:
                        type: string
                      role:
                        type: string
        '401':
          description: Invalid credentials

  /products:
    get:
      summary: List all products
      tags:
        - Products
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of products
    post:
      summary: Create a new product (Admin only)
      tags:
        - Products
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - price
              properties:
                sku:
                  type: string
                name:
                  type: string
                category_id:
                  type: integer
                price:
                  type: number
                cost_price:
                  type: number
                unit:
                  type: string
      responses:
        '201':
          description: Product created

  /products/{id}:
    get:
      summary: Get product by ID
      tags:
        - Products
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Product details
    put:
      summary: Update product (Admin only)
      tags:
        - Products
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Product updated
    delete:
      summary: Delete product (Admin only)
      tags:
        - Products
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Product deleted

  /products/search:
    get:
      summary: Search products
      tags:
        - Products
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Search results

  /inventory:
    get:
      summary: List all inventory
      tags:
        - Inventory
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of inventory items

  /inventory/{product_id}:
    get:
      summary: Get inventory by product ID
      tags:
        - Inventory
      security:
        - bearerAuth: []
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Inventory details

  /inventory/adjust:
    post:
      summary: Adjust inventory (Admin only)
      tags:
        - Inventory
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - product_id
                - delta
              properties:
                product_id:
                  type: integer
                delta:
                  type: integer
                reason:
                  type: string
      responses:
        '200':
          description: Inventory adjusted

  /sales:
    post:
      summary: Create a new sale
      tags:
        - Sales
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - items
                - paid_amount
              properties:
                items:
                  type: array
                  items:
                    type: object
                    required:
                      - product_id
                      - qty
                      - price
                    properties:
                      product_id:
                        type: integer
                      qty:
                        type: integer
                      price:
                        type: number
                      discount:
                        type: number
                paid_amount:
                  type: number
                payment_method:
                  type: string
      responses:
        '201':
          description: Sale created
    get:
      summary: List sales
      tags:
        - Sales
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of sales

  /sales/{id}:
    get:
      summary: Get sale by ID
      tags:
        - Sales
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Sale details

  /reports/sales:
    get:
      summary: Get sales report by date range
      tags:
        - Reports
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Sales report

  /reports/top-products:
    get:
      summary: Get top products report
      tags:
        - Reports
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Top products report

  /reports/stats:
    get:
      summary: Get sales statistics
      tags:
        - Reports
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Sales statistics

  /healthz:
    get:
      summary: Health check
      tags:
        - Health
      responses:
        '200':
          description: Service is healthy

security:
  - bearerAuth: []

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT